Add animation to the popup
