+++ index.html
@@ -292,6 +292,6 @@
       <p class="footer-email">javiervargasortega8683@gmail.com</p>
     </footer>
 
-    <script src="js/main.js" type="module"></script>
+    <script src="js/index.js" type="module"></script>
   </body>
 </html>
