+++ js/modules/projectsObj.js
@@ -1,0 +1,69 @@
+const projectObj = () => [
+  {
+    name: 'Conference Project',
+    title: 'Conference Project, Game DEV Conference',
+    description: "This is the capstone project for the 1st module at Microverse, the aim for this project is to show off and put into practice all the knowledge we've been learning throughout the 1st module.",
+    mobileSnapshot: 'images/capstone-project-mobile.jpg',
+    desktopSnapshot: 'images/Capstone-project.png',
+    technologies: ['HTML', 'CSS', 'JavaScript'],
+    linkLiveServer: 'https://javivy.github.io/capstone-1/index.html',
+    linkSource: 'https://github.com/Javivy/capstone-1',
+  },
+
+  {
+    name: 'Leaderboard Project',
+    title: 'Leaderboard Project, Making use of API',
+    description: "This is a leaderboard project, making use of an API from Microverse, working with HTML, CSS, JavaScript, Webpack, Linters, etc... On this project I practiced the use of API's, fetch, async-await, promises. Go check the live version and give it a try!",
+    mobileSnapshot: 'images/leaderboard-mobile.jpg',
+    desktopSnapshot: 'images/leaderboard-desktop.png',
+    technologies: ['HTML', 'CSS', 'JavaScript', 'Webpack', "API's"],
+    linkLiveServer: 'https://javivy.github.io/leaderboard/dist',
+    linkSource: 'https://github.com/Javivy/leaderboard',
+  },
+
+  {
+    name: 'Multi-Post stories Gain+Glory',
+    title: 'Project 3',
+    description: "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releaLorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the relea",
+    mobileSnapshot: 'images/capstone-project-mobile.jpg',
+    desktopSnapshot: 'images/Capstone-project.png',
+    technologies: ['Ruby on Rails', 'CSS', 'JavaScript', 'HTML'],
+    linkLiveServer: '#',
+    linkSource: '#',
+  },
+
+  {
+    name: 'Multi-Post stories Gain+Glory',
+    title: 'Project 4',
+    description: "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releaLorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the relea",
+    mobileSnapshot: 'images/capstone-project-mobile.jpg',
+    desktopSnapshot: 'images/Capstone-project.png',
+    technologies: ['Ruby on Rails', 'CSS', 'JavaScript', 'HTML'],
+    linkLiveServer: '#',
+    linkSource: '#',
+  },
+
+  {
+    name: 'Multi-Post stories Gain+Glory',
+    title: 'Project 5',
+    description: "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releaLorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the relea",
+    mobileSnapshot: 'images/capstone-project-mobile.jpg',
+    desktopSnapshot: 'images/Capstone-project.png',
+    technologies: ['Ruby on Rails', 'CSS', 'JavaScript', 'HTML'],
+    linkLiveServer: '#',
+    linkSource: '#',
+  },
+
+  {
+    name: 'Multi-Post stories Gain+Glory',
+    title: 'Project 6',
+    description: "Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releaLorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the relea",
+    mobileSnapshot: 'images/capstone-project-mobile.jpg',
+    desktopSnapshot: 'images/Capstone-project.png',
+    technologies: ['Ruby on Rails', 'CSS', 'JavaScript', 'HTML'],
+    linkLiveServer: '#',
+    linkSource: '#',
+  },
+];
+
+export default projectObj;
