+++ styles/styles.css
@@ -1012,3 +1012,15 @@ footer {
     transform: translateY(0);
   }
 }
+
+@keyframes slide-down {
+  from {
+    transform: translateY(0);
+    opacity: 1;
+  }
+
+  to {
+    transform: translateY(100%);
+    opacity: 0;
+  }
+}
